Removed global static schema Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
