+++ implementation/src/main/java/io/smallrye/graphql/execution/GraphQLProducer.java
@@ -1,0 +1,26 @@
+package io.smallrye.graphql.execution;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Produces;
+
+import graphql.schema.GraphQLSchema;
+
+/**
+ * Produces the GraphQL Schema
+ * 
+ * @author Phillip Kruger (phillip.kruger@redhat.com)
+ */
+@ApplicationScoped
+public class GraphQLProducer {
+
+    private GraphQLSchema graphQLSchema;
+
+    public void setGraphQLSchema(GraphQLSchema graphQLSchema) {
+        this.graphQLSchema = graphQLSchema;
+    }
+
+    @Produces
+    public GraphQLSchema getGraphQLSchema() {
+        return this.graphQLSchema;
+    }
+}
