+++ tck/pom.xml
@@ -16,19 +16,19 @@
         <dependency>
             <groupId>org.eclipse.microprofile.graphql</groupId>
             <artifactId>microprofile-graphql-tck</artifactId>
-            <!--<scope>test</scope>-->
+            <scope>test</scope>
         </dependency>
         
         <dependency>
             <groupId>io.smallrye</groupId>
             <artifactId>smallrye-graphql</artifactId>
-            <!--<scope>test</scope> This breaks the coverage ?!? --> 
+            <scope>test</scope>
         </dependency>
         
         <dependency>
             <groupId>io.smallrye</groupId>
             <artifactId>smallrye-graphql-servlet</artifactId>
-            <!--<scope>test</scope> This breaks the coverage ?!? --> 
+            <scope>test</scope>
         </dependency>
         
         <dependency>
@@ -52,6 +52,7 @@
                 <artifactId>maven-surefire-plugin</artifactId>
                 <configuration>
                     <dependenciesToScan>
+                        <!--<dependency>org.eclipse.microprofile.graphql:microprofile-graphql-server-tck</dependency>-->
                         <dependency>org.eclipse.microprofile.graphql:microprofile-graphql-tck</dependency>
                     </dependenciesToScan>
                 </configuration>
